feat: 좋아요 처리, 집계 리팩토링 및 카프카 이벤트 발행(Consumer)
