+++ apps/commerce-api/src/main/java/com/loopers/application/like/event/LikeEventHandler.java
@@ -1,8 +1,7 @@
 package com.loopers.application.like.event;
 
+import com.loopers.application.like.LikeProcessingService;
 import com.loopers.domain.like.LikeType;
-import com.loopers.domain.product.ProductCacheRepository;
-import com.loopers.domain.product.ProductService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.orm.ObjectOptimisticLockingFailureException;
@@ -23,72 +22,26 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class LikeEventHandler {
 
-    private final ProductService productService;
-    private final ProductCacheRepository productCacheRepository;
+    private final LikeProcessingService likeProcessingService;
 
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
     @EventListener
     @Async
     public void handleLikeAdded(LikeAddedEvent event) {
-        log.info("좋아요 추가 이벤트 처리 시작 - userId: {}, targetId: {}, likeType: {}", 
-                event.userId(), event.targetId(), event.likeType());
+        log.info("좋아요 추가 이벤트 처리 시작 - userId: {}, targetId: {}, likeType: {}", event.userId(), event.targetId(), event.likeType());
 
-        if (event.likeType() == LikeType.PRODUCT) {
-            processProductLikeIncrease(event.targetId());
-        }
+        likeProcessingService.processLikeAdded(event);
 
         log.info("좋아요 추가 집계 처리 완료 - targetId: {}", event.targetId());
     }
 
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
     @EventListener
     @Async
     public void handleLikeRemoved(LikeRemovedEvent event) {
-        log.info("좋아요 제거 이벤트 처리 시작 - userId: {}, targetId: {}, likeType: {}", 
-                event.userId(), event.targetId(), event.likeType());
+        log.info("좋아요 제거 이벤트 처리 시작 - userId: {}, targetId: {}, likeType: {}", event.userId(), event.targetId(), event.likeType());
 
-        if (event.likeType() == LikeType.PRODUCT) {
-            processProductLikeDecrease(event.targetId());
-        }
+        likeProcessingService.processLikeRemoved(event);
 
         log.info("좋아요 제거 집계 처리 완료 - targetId: {}", event.targetId());
     }
 
-    @Retryable(
-            value = { ObjectOptimisticLockingFailureException.class },
-            maxAttempts = 50,
-            backoff = @Backoff(delay = 10, multiplier = 1.5, maxDelay = 1000)
-    )
-    private void processProductLikeIncrease(Long productId) {
-        try {
-            productService.increaseLikeCount(productId);
-            
-            // Redis 캐시 삭제(상품의 총 좋아요 수)
-            productCacheRepository.evictProductDetail(productId);
-            
-            log.info("상품 좋아요 수 증가 완료 - productId: {}", productId);
-        } catch (Exception e) {
-            log.error("상품 좋아요 수 증가 실패 - productId: {}, error: {}", productId, e.getMessage(), e);
-            throw e;
-        }
-    }
-
-    @Retryable(
-            value = { ObjectOptimisticLockingFailureException.class },
-            maxAttempts = 50,
-            backoff = @Backoff(delay = 10, multiplier = 1.5, maxDelay = 1000)
-    )
-    private void processProductLikeDecrease(Long productId) {
-        try {
-            productService.decreaseLikeCount(productId);
-            
-            // Redis 캐시 삭제(상품의 총 좋아요 수)
-            productCacheRepository.evictProductDetail(productId);
-            
-            log.info("상품 좋아요 수 감소 완료 - productId: {}", productId);
-        } catch (Exception e) {
-            log.error("상품 좋아요 수 감소 실패 - productId: {}, error: {}", productId, e.getMessage(), e);
-            throw e;
-        }
-    }
 }
